backend/pyproject.toml
[project]
name = "moderation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "connexion[flask,swagger-ui,uvicorn]>=3.2.0",
    "kafka-python>=2.1.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.3",
    "sqlalchemy>=2.0.40",
    "dependency-injector>=4.46.0",
    "pillow>=11.2.1",
    "python-json-logger>=3.3.0",
    "pyjwt>=2.10.1",
    "redis>=6.1.0",
    "faker>=37.3.0",
    "pytest>=8.3.5",
    "types-redis>=4.6.0.20241004",
]
